analyzer/requirements.txt
python-binance==1.0.19
asyncpg==0.29.0
aiohttp==3.9.3
pandas==2.2.0
numpy==1.26.4
pandas-ta==0.3.14b
scikit-learn==1.4.0
websockets==12.0
